Moving Client to a separate file.
