+++ src/action.ts
@@ -1,38 +1,10 @@
 import { init, last } from "ramda";
-import type { Socket } from "socket.io";
 import { z } from "zod";
 import { ActionNoAckDef, ActionWithAckDef } from "./actions-factory";
-import { Broadcaster, EmissionMap, Emitter, RoomService } from "./emission";
+import { Client } from "./client";
+import { EmissionMap, RoomService } from "./emission";
 import { AbstractLogger } from "./logger";
 import { RemoteClient } from "./remote-client";
-
-export interface Client<E extends EmissionMap> {
-  /** @alias Socket.connected */
-  isConnected: () => boolean;
-  /** @alias Socket.id */
-  id: Socket["id"];
-  /** @desc Returns the list of the rooms the client in */
-  getRooms: () => string[];
-  /**
-   * @desc Sends a new event to the client (this is not acknowledgement)
-   * @throws z.ZodError on validation
-   * @throws Error on ack timeout
-   * */
-  emit: Emitter<E>;
-  /**
-   * @desc Emits to others
-   * @throws z.ZodError on validation
-   * @throws Error on ack timeout
-   * */
-  broadcast: Broadcaster<E>;
-  /** @desc Returns the client metadata according to the specified type or empty object */
-  getData: <D extends object>() => Readonly<Partial<D>>;
-  /**
-   * @desc Sets the client metadata according to the specified type
-   * @throws z.ZodError on validation
-   * */
-  setData: <D extends object>(value: D) => void;
-}
 
 export interface HandlingFeatures<E extends EmissionMap> {
   logger: AbstractLogger;
